// emmaus_ui/src/app/services/models/creat-staff-profile-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreatStaffProfileRequest {
  bio?: string;
  department?: string;
  description?: string;
  displayOrder?: number;
  email?: string;
  firstName?: string;
  lastName?: string;
  linkedinUrl?: string;
  phone?: string;
  position?: string;
  profilePic?: string;
}